+++ packages/react-components/react-dialog/src/components/DialogActions/DialogActions.test.tsx
@@ -8,7 +8,6 @@ describe('DialogActions', () => {
   isConformant<DialogActionsProps>({
     Component: DialogActions,
     displayName: 'DialogActions',
-    disabledTests: ['component-has-static-classname-exported'],
   });
 
   // TODO add more tests here, and create visual regression tests in /apps/vr-tests