+++ packages/react-components/react-menu/src/components/MenuSplitGroup/MenuSplitGroup.test.tsx
@@ -6,7 +6,6 @@ import { isConformant } from '../../common/isConformant';
 describe('MenuSplitGroup', () => {
   isConformant({
     Component: MenuSplitGroup,
-    disabledTests: ['component-has-static-classname-exported'],
     displayName: 'MenuSplitGroup',
   });