+++ packages/react-components/react-table/src/components/TableBody/TableBody.types.ts
@@ -1,0 +1,15 @@
+import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
+
+export type TableBodySlots = {
+  root: Slot<'tbody', 'div'>;
+};
+
+/**
+ * TableBody Props
+ */
+export type TableBodyProps = ComponentProps<TableBodySlots>;
+
+/**
+ * State used in rendering TableBody
+ */
+export type TableBodyState = ComponentState<TableBodySlots>;