+++ packages/react-components/react-table/src/components/TableRow/TableRow.types.ts
@@ -1,0 +1,16 @@
+import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
+import { TableContextValue } from '../../contexts/types';
+
+export type TableRowSlots = {
+  root: Slot<'tr', 'div'>;
+};
+
+/**
+ * TableRow Props
+ */
+export type TableRowProps = ComponentProps<TableRowSlots> & {};
+
+/**
+ * State used in rendering TableRow
+ */
+export type TableRowState = ComponentState<TableRowSlots> & { size: TableContextValue['size'] };