+++ packages/react-components/react-table/src/index.ts
@@ -1,2 +1,4 @@
-// TODO: replace with real exports
-export {};
+export { TableCell, tableCellClassNames, tableCellClassName } from './TableCell';
+export { TableRow, tableRowClassNames, tableRowClassName } from './TableRow';
+export { TableBody, tableBodyClassName, tableBodyClassNames } from './TableBody';
+export { Table, tableClassName, tableClassNames } from './Table';