+++ packages/react-components/react-text/src/components/Title1/Title1.test.tsx
@@ -7,7 +7,6 @@ describe('Title1', () => {
   isConformant({
     Component: Title1,
     displayName: 'Title1',
-    disabledTests: ['component-has-static-classname', 'component-has-static-classname-exported'],
   });
 
   it('renders a default state', () => {